Todo list columns are locked in position
